refactor(location-table): extract table
